Removed casting as i32

+++ src/lib.rs
@@ -218,7 +218,7 @@ pub fn fseek( fd: i32, operation: SeekFrom ) -> Result<i64,Error> {
             dist = x;
         }
     };
-    let ret = unsafe{ lseek( fd, dist as i32, seek_op )};
+    let ret = unsafe{ lseek( fd, dist, seek_op )};
     if ret == 0 {
         Ok(ret as i64)
     } else {
